Use cogs:Jobs and integrate with Job controller
We now start from a task created by the job-controller with an input
data container. Some functions have been removed or rewritten because of
this. We also write different states to the tasks status to make sure
the job-controller can pick up on the progress.
All constants have been moved to a separate file.