Start task based on logical file URI
Previously, a physical file URI was saved as the input container for
this service, that has now changed to a logical file URI and these
changes make sure the correct URI is passed.
Also, the physical file also gets a new related logical file, and not
the remote data object originally from the download-url-service, because
this really is a different file after all. There is still a link to the
submission, but to the logical file instead of the physical file.